Try out hiding the original patches from the model when providing a reflection.
I'm finding this has middling results. The reflection seems to trigger worse
performance as much as it improves it. For instance, for the affine-cipher, the
tests are clearly intended to show that numbers should be carried through
unchanged, but the reflection has the model try to implement a larger space.